[Chore] #85 Breed 도메인 분리
