+++ src/main/java/com/kuit/findyou/domain/breed/dto/BreedResponseDTO.java
@@ -1,10 +1,9 @@
-package com.kuit.findyou.domain.report.dto;
+package com.kuit.findyou.domain.breed.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
