+++ src/main/java/com/kuit/findyou/domain/breed/dto/BreedValidateResponseDTO.java
@@ -1,8 +1,7 @@
-package com.kuit.findyou.domain.report.dto;
+package com.kuit.findyou.domain.breed.dto;
 
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
