+++ src/main/java/com/kuit/findyou/domain/breed/service/GetAllBreedsService.java
@@ -1,7 +1,7 @@
-package com.kuit.findyou.domain.report.service;
+package com.kuit.findyou.domain.breed.service;
 
-import com.kuit.findyou.domain.report.dto.BreedResponseDTO;
-import com.kuit.findyou.domain.report.repository.BreedRepository;
+import com.kuit.findyou.domain.breed.dto.BreedResponseDTO;
+import com.kuit.findyou.domain.breed.repository.BreedRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
