+++ src/main/java/com/kuit/findyou/domain/report/service/WitnessReportPostService.java
@@ -1,5 +1,6 @@
 package com.kuit.findyou.domain.report.service;
 
+import com.kuit.findyou.domain.breed.model.Breed;
 import com.kuit.findyou.domain.user.model.User;
 import com.kuit.findyou.domain.auth.repository.UserRepository;
 import com.kuit.findyou.domain.home.dto.ReportTag;
@@ -8,7 +9,7 @@ import com.kuit.findyou.domain.report.dto.WitnessReportDTO;
 import com.kuit.findyou.domain.report.exception.ReportCreationException;
 import com.kuit.findyou.domain.report.model.*;
 import com.kuit.findyou.domain.report.repository.AnimalFeatureRepository;
-import com.kuit.findyou.domain.report.repository.BreedRepository;
+import com.kuit.findyou.domain.breed.repository.BreedRepository;
 import com.kuit.findyou.domain.image.repository.ImageRepository;
 import com.kuit.findyou.domain.report.repository.ReportRepository;
 import com.kuit.findyou.global.common.response.status.BaseExceptionResponseStatus;
